+++ pkg/consts/consts.go
@@ -119,6 +119,14 @@ const (
 
 	// CAFieldSecret is the field name in Kubernetes secret - WebhookCertificateConfigSecretName.
 	CAFieldSecret = "ca"
+
+	// KongClusterCertVolume is the name of the volume that holds the certificate the enables
+	// communication between Kong and Konnect.
+	KongClusterCertVolume = "kong-cluster-cert"
+
+	// KongClusterCertVolumeMountPath holds the path where the Kong Cluster certificate
+	// volume will be mounted.
+	KongClusterCertVolumeMountPath = "/etc/secrets/kong-cluster-cert"
 )
 
 // -----------------------------------------------------------------------------